requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
django-cors-headers>=4.0
python-dotenv>=1.0

# Optional / production
psycopg2-binary>=2.9
whitenoise>=6.0
gunicorn>=20.1
